+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:53605/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:53830/jest?","mongoDBName":"jest"}
