+++ src/data/use-cases/add-account/db-add-account-protocols.ts
@@ -1,4 +1,4 @@
 export * from '../../../domain/models/account'
 export * from '../../../domain/use-cases/add-account'
-export * from '../../protocols/encrypter'
-export * from '../../protocols/add-account-repository'
+export * from '../../protocols/criptography/encrypter'
+export * from '../../protocols/db/add-account-repository'
