+++ src/data/use-cases/authentication/db-authentication.ts
@@ -1,5 +1,5 @@
 import { Authentication, AuthenticationModel } from '../../../domain/use-cases/authentication'
-import { LoadAccountByEmailRepository } from '../../protocols/load-account-by-email-repository'
+import { LoadAccountByEmailRepository } from '../../protocols/db/load-account-by-email-repository'
 
 export class DbAuthentication implements Authentication {
   private readonly loadAccountByEmailRepository: LoadAccountByEmailRepository
